routing logic entirely moved to src/components/layout.jsx
