+++ src/pages/404.js
@@ -1,20 +1,13 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
-import { Redirect, Router } from '@reach/router';
-import { navigate } from 'gatsby';
-import { useTranslation } from 'react-i18next';
+// import { Redirect, Router } from '@reach/router';
 
-import i18next from '../components/i18n';
+import Layout from '../components/layout';
 
 function NotFound() {
-  const { i18n } = useTranslation();
-  useEffect(() => {
-    navigate(`/${i18n.language}/404`);
-  }, []);
-
   return (
     // eslint-disable-next-line react/jsx-filename-extension
-    <></>
+    <Layout />
     // // <Router>
     // //   <Redirect from="/" to={`/${i18n.language}/404`} noThrow />
     // // </Router>
