+++ src/templates/landing-page.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
 import { GridContainer } from '@trussworks/react-uswds';
 import { graphql } from 'gatsby';
@@ -24,34 +24,8 @@ import '@fontsource/space-mono';
 
 import '../styles/index.scss';
 
-function Landing({ data, location }) {
+function Landing({ data }) {
   const { t, i18n } = useTranslation();
-
-  // TODO: try to use LanguageDetector
-  useEffect(() => {
-    const path = location.pathname;
-    const lang = path.split('/')[1];
-    const route = path
-      .split('/')
-      .slice(2)
-      .filter((v) => v !== '')
-      .join('/');
-    if (lang.length === 0 && route.length === 0) {
-      // empty path
-      // not currently being used
-      // navigate(`/${i18n.language}/`);
-    } else if (lang.length !== 0 && route.length === 0) {
-      // only language given
-      if (lang !== i18n.language) i18n.changeLanguage(lang);
-    } else if (lang.length === 0 && route.length !== 0) {
-      // only route given
-      // not currently being used
-      // navigate(`${i18n.language}/${route}`);
-    } else if (lang !== i18n.language) {
-      // both language and route given
-      i18n.changeLanguage(lang);
-    }
-  }, [location.pathname]);
 
   const { markdownRemark } = data;
   const { frontmatter } = markdownRemark;
@@ -92,7 +66,6 @@ function Landing({ data, location }) {
 
 Landing.propTypes = {
   data: PropTypes.node.isRequired,
-  location: PropTypes.node.isRequired,
 };
 
 export const pageQuery = graphql`
